+++ src/main/java/ifce/viviservice/entity/Curriculo.java
@@ -19,7 +19,7 @@ public class Curriculo {
     @Id
     @Column(name = "CODIGO")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer codigo;
+    private Long codigo;
 
     @Column(name = "DESCRICAO")
     private String descricao;
