+++ src/main/java/ifce/viviservice/repository/AdministradorRepository.java
@@ -3,6 +3,6 @@ package ifce.viviservice.repository;
 import ifce.viviservice.entity.Administrador;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AdministradorRepository extends JpaRepository<Administrador, Integer> {
+public interface AdministradorRepository extends JpaRepository<Administrador, Long> {
 
 }
