+++ src/main/java/ifce/viviservice/repository/AutenticacaoRepository.java
@@ -3,6 +3,6 @@ package ifce.viviservice.repository;
 import ifce.viviservice.entity.Autenticacao;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AutenticacaoRepository extends JpaRepository<Autenticacao, Integer> {
+public interface AutenticacaoRepository extends JpaRepository<Autenticacao, Long> {
 
 }
