+++ src/main/java/ifce/viviservice/repository/CampusRepository.java
@@ -3,6 +3,6 @@ package ifce.viviservice.repository;
 import ifce.viviservice.entity.Campus;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CampusRepository extends JpaRepository<Campus, Integer> {
+public interface CampusRepository extends JpaRepository<Campus, Long> {
 
 }
