+++ src/main/java/ifce/viviservice/repository/ExperienciaRepository.java
@@ -3,6 +3,6 @@ package ifce.viviservice.repository;
 import ifce.viviservice.entity.Experiencia;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ExperienciaRepository extends JpaRepository<Experiencia, Integer> {
+public interface ExperienciaRepository extends JpaRepository<Experiencia, Long> {
 
 }
